Move Log4j Logger class extension to bootstrap.rb
